// frontend/src/api/models/updateStatoRepairRequestStato.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Gestionale Negozio API
 * API del gestionale negozio
 * OpenAPI spec version: 1.0.0
 */

export type UpdateStatoRepairRequestStato = typeof UpdateStatoRepairRequestStato[keyof typeof UpdateStatoRepairRequestStato];


export const UpdateStatoRepairRequestStato = {
  NUOVO: 'NUOVO',
  IN_CORSO: 'IN_CORSO',
  PRONTO: 'PRONTO',
  CONSEGNATO: 'CONSEGNATO',
} as const;