// frontend/src/api/models/updateStatoRepairRequestStatoRiparazione.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Gestionale Negozio API
 * API del gestionale negozio
 * OpenAPI spec version: 1.0.0
 */

export type UpdateStatoRepairRequestStatoRiparazione = typeof UpdateStatoRepairRequestStatoRiparazione[keyof typeof UpdateStatoRepairRequestStatoRiparazione];


export const UpdateStatoRepairRequestStatoRiparazione = {
  ACCETTATO: 'ACCETTATO',
  ANALISI_IN_CORSO: 'ANALISI_IN_CORSO',
  RIPARAZIONE_IN_CORSO: 'RIPARAZIONE_IN_CORSO',
  ATTESA_PEZZI_DI_RICAMBIO: 'ATTESA_PEZZI_DI_RICAMBIO',
  IN_ATTESA_DI_PREVENTIVO: 'IN_ATTESA_DI_PREVENTIVO',
  PREVENTIVO_NON_ACCETTATO: 'PREVENTIVO_NON_ACCETTATO',
  RIPARAZIONE_CONCLUSA: 'RIPARAZIONE_CONCLUSA',
  DISPOSITIVO_NON_RIPARABILE: 'DISPOSITIVO_NON_RIPARABILE',
} as const;